# openapi/raibitserver.yaml
openapi: 3.1.0
info:
  title: RAIBITSERVER API
  version: 0.2.0
  description: Canonical desired-state contract shared by NestJS API, prototype handler, API client, CLI, and local E2E.
servers:
  - url: http://localhost:3000/api
  - url: https://api.raibitserver.app
security:
  - bearerAuth: []
paths:
...
  /snapshot: { get: { summary: Return a bounded administrative overview, parameters: [ { $ref: '#/components/parameters/limit' } ], responses: { '200': { description: Bounded users, quotas, and recent audit logs }, '403': { description: Administrator permission required } } } }
...
  /projects: { get: { summary: List visible projects, parameters: [ { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated projects } } }, post: { summary: Create project using subject organization, responses: { '201': { description: Created } } } }
  /projects/{projectId}: { get: { summary: Get project, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '200': { description: Project } } }, patch: { summary: Update project, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '200': { description: Updated } } }, delete: { summary: Delete project, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '204': { description: Deleted } } } }
  /projects/{projectId}/overview: { get: { summary: Get bounded project console overview, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '200': { description: Project, services, deployments, resources, and recent activity in project scope }, '403': { description: Project access denied }, '404': { description: Project not found } } } }
  /projects/{projectId}/services: { get: { summary: List services, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '200': { description: Services } } }, post: { summary: Create service desired state, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '201': { description: Created } } } }
  /projects/{projectId}/services/{serviceId}/deployments: { get: { summary: List deployments for a service in a project, parameters: [ { $ref: '#/components/parameters/projectId' }, { $ref: '#/components/parameters/serviceId' } ], responses: { '200': { description: Deployments } } }, post: { summary: Queue production/manual/preview deployment in project scope, parameters: [ { $ref: '#/components/parameters/projectId' }, { $ref: '#/components/parameters/serviceId' } ], responses: { '202': { description: Deployment and workflow job queued with preview URL/workload metadata when requested } } } }
  /services/{serviceId}: { get: { summary: Get service, parameters: [ { $ref: '#/components/parameters/serviceId' } ], responses: { '200': { description: Service } } }, patch: { summary: Update service, parameters: [ { $ref: '#/components/parameters/serviceId' } ], responses: { '200': { description: Updated } } }, delete: { summary: Delete service, parameters: [ { $ref: '#/components/parameters/serviceId' } ], responses: { '204': { description: Deleted } } } }
  /services/{serviceId}/deployments: { get: { summary: List deployments, parameters: [ { $ref: '#/components/parameters/serviceId' }, { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated deployments } } }, post: { summary: Queue production/manual/preview deployment, parameters: [ { $ref: '#/components/parameters/serviceId' } ], responses: { '202': { description: Deployment and workflow job queued } } } }
  /deployments/{deploymentId}: { get: { summary: Get deployment, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], responses: { '200': { description: Deployment } } } }
  /deployments/{deploymentId}/status: { patch: { summary: Update deployment status/image/error fields, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], responses: { '200': { description: Updated deployment } } }, post: { summary: Update deployment status/image/error fields from HTML forms, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], responses: { '200': { description: Updated deployment } } } }
  /deployments/{deploymentId}/cancel: { post: { summary: Cancel a deployment before runtime reconciliation starts, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], responses: { '200': { description: Deployment cancelled atomically }, '409': { description: Deployment can no longer be cancelled safely } } } }
  /deployments/{deploymentId}/rollback: { post: { summary: Roll back to previous READY image after explicit confirmation, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [confirmed], properties: { confirmed: { type: boolean, const: true }, previousDeploymentId: { type: string }, imageUrl: { type: string } } } } } }, responses: { '202': { description: Rollback queued }, '400': { description: Explicit confirmation is required } } } }
  /deployments/{deploymentId}/logs: { get: { summary: List build logs, parameters: [ { $ref: '#/components/parameters/deploymentId' }, { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated build logs } } } }
  /deployments/{deploymentId}/events: { get: { summary: List deployment events, parameters: [ { $ref: '#/components/parameters/deploymentId' }, { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated deployment events } } } }
  /deployments/{deploymentId}/stream: { get: { summary: Stream deployment logs and events as Server-Sent Events, parameters: [ { $ref: '#/components/parameters/deploymentId' } ], responses: { '200': { description: SSE deployment activity stream } } } }
...
  /services/{serviceId}/logs: { get: { summary: List runtime logs, parameters: [ { $ref: '#/components/parameters/serviceId' }, { $ref: '#/components/parameters/limit' }, { $ref: '#/components/parameters/cursor' } ], responses: { '200': { description: Keyset-paginated runtime logs } } } }
  /services/{serviceId}/logs/stream: { get: { summary: Stream service runtime logs as Server-Sent Events, parameters: [ { $ref: '#/components/parameters/serviceId' } ], responses: { '200': { description: SSE runtime log stream } } } }
  /projects/{projectId}/resources: { get: { summary: List resources, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '200': { description: Resources } } }, post: { summary: Create managed resource, parameters: [ { $ref: '#/components/parameters/projectId' } ], responses: { '201': { description: Created } } } }
  /resources/{resourceId}: { get: { summary: Get resource, parameters: [ { $ref: '#/components/parameters/resourceId' } ], responses: { '200': { description: Resource } } }, patch: { summary: Update resource, parameters: [ { $ref: '#/components/parameters/resourceId' } ], responses: { '200': { description: Updated } } }, delete: { summary: Delete resource and provider-owned secrets, parameters: [ { $ref: '#/components/parameters/resourceId' } ], responses: { '200': { description: Deleted } } } }
  /resources/{resourceId}/attach: { post: { summary: Attach provider-owned resource env to service secrets, parameters: [ { $ref: '#/components/parameters/resourceId' } ], responses: { '200': { description: Attached } } } }
  /resources/{resourceId}/provision: { post: { summary: Build/apply resource provider plan and provider-owned connection secrets, parameters: [ { $ref: '#/components/parameters/resourceId' } ], responses: { '202': { description: Provision result } } } }
...
  /admin/users/{userId}/approve: { post: { summary: Approve user, parameters: [ { name: userId, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: Approved } } } }
  /admin/users/{userId}/reject: { post: { summary: Reject user after explicit confirmation, parameters: [ { name: userId, in: path, required: true, schema: { type: string } } ], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [confirmed], properties: { confirmed: { type: boolean, const: true } } } } } }, responses: { '200': { description: Rejected }, '400': { description: Explicit confirmation is required } } } }
  /admin/users/{userId}/quota: { patch: { summary: Update user quota, parameters: [ { name: userId, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: Updated } } } }
...
components:
  parameters:
...
    projectId: { name: projectId, in: path, required: true, schema: { type: string } }
    serviceId: { name: serviceId, in: path, required: true, schema: { type: string } }
    deploymentId: { name: deploymentId, in: path, required: true, schema: { type: string } }
    resourceId: { name: resourceId, in: path, required: true, schema: { type: string } }
...
    limit: { name: limit, in: query, required: false, description: Maximum rows per page, schema: { type: integer, minimum: 1, maximum: 1000, default: 200 } }
    cursor: { name: cursor, in: query, required: false, description: Opaque versioned keyset cursor returned by the previous page, schema: { type: string, maxLength: 1024 } }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...